+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # rollup-plugin-typescript changelog
 
+## 0.2.1
+* Author the plugin in TypeScript
+* Log diagnostics encountered during transpilation
+* Generate source maps per default
+
 ## 0.2.0
 * Use (_prerelease version of_) TypeScript 1.7.0 to generate ES5 while preserving ES2015 imports for efficient bundling.